+++ client/src/components/Pages/auth/CheckCode/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CheckCode";
