+++ server/src/workflows/users/sendCode/index.ts
@@ -1,0 +1,1 @@
+export * from "./sendCode";
